+++ lib/memstream.c
@@ -101,7 +101,7 @@ unsigned int mread32(memstream *s){
 	if(!s||s->size-s->current<4)return 0;
 	const unsigned char *x=(const unsigned char*)s->p+s->current;
 	s->current+=4;
-	return x[0]|(x[1]<<8)|(x[2]<<16)|(x[3]<<24);
+	return x[0]|(x[1]<<8)|(x[2]<<16)|((unsigned int)x[3]<<24);
 }
 
 unsigned short mread16(memstream *s){
